+++ R/isForkedChild.R
@@ -10,9 +10,10 @@
 #'
 #' @export
 isForkedChild <- local({
-  isChild <- NULL
+  .cache <- NULL  ## Cannot name just 'isChild', bc R CMD check/codetools
   
   function() {
+    isChild <- .cache
     if (is.null(isChild)) {
       if (supportsMulticore()) {
         ## Asked for parallel:::isChild() to be exported /HB 2021-11-04
@@ -21,8 +22,8 @@ isForkedChild <- local({
       } else {
         isChild <- function() FALSE
       }
+      .cache <<- isChild
     }
-    
     isChild()
   }
 })